reorganize code to make it easier to run coverage
