+++ backend/program/indexers/__init__.py
@@ -1,0 +1,1 @@
+from .trakt import TraktIndexer
